nextjs/works/requirements.txt
boto3==1.34.88
botocore==1.34.88
doctr==1.9.0
fitz==0.0.1.dev2
fuzzywuzzy==0.18.0
numpy==1.23.5
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
pandas==1.5.3
Pillow==10.3.0
protobuf==5.26.1
pytesseract==0.3.10
python_doctr==0.8.1
streamlit==1.32.2
ultralytics==8.1.47
